Server: Fix posting stuff into the inbox when the ID already exists in the database
Basically, the issue is that when saving a thing in the database that already exists, the existing entity is changed. If the code adds the item to the collection before it's commited into the database, it might use the wrong object, which implicitly adds it into the database. Causing duplicate keys.