Server: Make internal and external federation make more sense and align
So this is a twofold change. First, the server uses an internal mechanism to federate to other people on the same server. Secondly, like/follow/announce effects now only occur when the owner of that object is notified.

Why? Because this aligns the effects with the case where everyone would run their own server. Also it makes sense that a user would need to know about likes on their own object.